+++ rapydo/flask_ext/flask_celery/worker.py
