+++ rapydo/flask_ext/flask_neo4j.py
@@ -5,7 +5,7 @@
 import socket
 import neo4j
 from neomodel import db, config
-from flask_ext import BaseExtension, get_logger
+from rapydo.flask_ext import BaseExtension, get_logger
 from rapydo.utils.logs import re_obscure_pattern
 
 log = get_logger(__name__)
