+++ rapydo/services/detect.py
@@ -10,8 +10,7 @@ Note: docker links and automatic variables removed as unsafe with compose V3
 
 import os
 # from functools import lru_cache
-from rapydo.confs import CORE_CONFIG_PATH
-from rapydo.confs import CUSTOM_PACKAGE
+from rapydo.confs import CORE_CONFIG_PATH, BACKEND_PACKAGE, CUSTOM_PACKAGE
 from rapydo.utils.meta import Meta
 from rapydo.utils.myyaml import load_yaml_file
 
@@ -125,7 +124,9 @@ class Detector(object):
 
         # Try inside our extensions
         module = self.meta.get_module_from_string(
-            modulestring='flask_ext' + flaskext, exit_on_fail=True)
+            modulestring=BACKEND_PACKAGE + '.flask_ext' + flaskext,
+            exit_on_fail=True
+        )
         if module is None:
             log.critical_exit("Missing %s for %s" % (flaskext, service))
